requirements.txt
requests
kafka-python
elasticsearch
junos-eznc
